+++ src/raytracer/objects/cylinder/Cylinder.cpp
@@ -31,14 +31,14 @@ void raytracer::Cylinder::parse(const raytracer::Raytracer& raytracer, const lib
     // Other settings
     if (node.exists("dimension")) {
         const libconfig::Setting& dim = node["dimension"];
-        descriptor.dimension.y = dim[0];
+        descriptor.dimension.y = (int)dim[0];
     } else {
         descriptor.infinite = true;
     }
 
     if (!node.exists("radius"))
         throw utils::exception::CustomException(utils::exception::Error, utils::exception::Code::Parser, "The radius field isn't defined for the object");
-    descriptor.radius = node["radius"];
+    descriptor.radius = (double)node["radius"];
 
     // Set the descriptor
     this->setShapeDescriptor(descriptor);
